updating FileManager and p2p interaction

+++ torrent/src/main/java/client/FileDescr.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class FileDescr {
     private final int id;
     private final String name;
-    private final long size;
+    private final int size;
 }
